requirements.txt
python-dotenv
pandas 
numpy
matplotlib
scikit-learn
pymongo[srv]
certifi
#-e .